convert server to typescript
