+++ package.json
@@ -99,9 +99,11 @@
     "@testing-library/react-hooks": "^3.2.1",
     "@types/enzyme": "^3.10.4",
     "@types/enzyme-adapter-react-16": "^1.0.5",
+    "@types/inflection": "^1.5.28",
     "@types/react-facebook-login": "^4.1.1",
     "@types/react-grid-layout": "^0.17.1",
     "@types/react-sound": "^1.2.2",
+    "clean-webpack-plugin": "^3.0.0",
     "cypress": "^3.8.3",
     "cypress-multi-reporters": "^1.2.3",
     "enzyme": "^3.11.0",
@@ -117,8 +119,12 @@
     "pixelmatch": "^5.1.0",
     "puppeteer": "^2.1.1",
     "react-test-renderer": "^16.12.0",
+    "source-map-loader": "^0.2.4",
     "start-server-and-test": "^1.10.8",
-    "ts-jest": "^24.3.0"
+    "ts-jest": "^24.3.0",
+    "ts-loader": "^6.2.2",
+    "webpack-node-externals": "^1.7.2",
+    "webpack-shell-plugin": "^0.5.0"
   },
   "lint-staged": {
     "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": [
@@ -132,7 +138,7 @@
     "build": "node scripts/build.js",
     "test": "jest",
     "lint": "eslint src/**/*.{ts,tsx}",
-    "server": "node server.js",
+    "server": "node scripts/start-server.js",
     "test:ci": "jest --ci --runInBand --reporters=default --reporters=jest-junit --coverage",
     "cypress:open": "cypress open",
     "cypress:run": "cypress run --browser chrome",
